Enabling the check that key names are UTF-8 compliant.
Note that this check means that all ODB key names will be checked for UTF-8 compliance when the ODB is first accessed by any program.  Any non-UTF8 compliant key names will be replaced by a string of the pointer to the key.

This behaviour (automatically fixing ODB names) can be disabled by setting an environment variable

MIDAS_INVALID_STRING_IS_OK

(it doesn't matter what variable is set to; it just needs to be set.)